refactor: Migrates a few not-so-shared files from the shared folder to appropriate directories.
