+++ src/app/infrastructure/auth/jwt/constants.ts
@@ -1,3 +1,2 @@
-export const USER = 'user';
 export const ACCESS_TOKEN = 'access-token';
 export const REFRESH_TOKEN = 'refresh-token';
