+++ src/app/modules/learning/knowledge-component/learning-objects/assessment-items/arrays.ts
